+++ src/app/layout.tsx
@@ -3,6 +3,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import { CartProvider } from "@/lib/context/CartContext";
 import { Toaster } from 'sonner'
+import ProgressBar from "@/components/ui/ProgressBar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -20,6 +21,7 @@ export default function RootLayout({
     <html lang="en">
       <body className={inter.className}>
         <CartProvider>
+          <ProgressBar />
           {children}
           <Toaster position="top-center" />
         </CartProvider>